filter data list to int list

+++ COT.Data/src/DataService/CoreLibrary/Interfaces/IFilterData.cs
@@ -4,6 +4,6 @@ namespace CoreLibrary.Interfaces
 {
     public interface IFilterData
     {
-        IList<double> Filter(string inputData, string symbol);
+        IList<int> Filter(string inputData, string symbol);
     }
 }
